make some cleaning stuff better
