+++ src/main/java/amidst/fragment/AvailableFragmentCache.java
@@ -1,14 +1,15 @@
 package amidst.fragment;
 
 import java.lang.ref.SoftReference;
+import java.util.Iterator;
 import java.util.concurrent.ConcurrentLinkedQueue;
 
 import amidst.documentation.AmidstThread;
 import amidst.documentation.CalledOnlyBy;
-import amidst.documentation.ThreadSafe;
+import amidst.documentation.NotThreadSafe;
 import amidst.fragment.constructor.FragmentConstructor;
 
-@ThreadSafe
+@NotThreadSafe
 public class AvailableFragmentCache {
 	private final ConcurrentLinkedQueue<SoftReference<Fragment>> cache = new ConcurrentLinkedQueue<>();
 	
@@ -37,16 +38,7 @@ public class AvailableFragmentCache {
 	@CalledOnlyBy(AmidstThread.EDT)
 	private Fragment poll() {
 		SoftReference<Fragment> ref = (SoftReference<Fragment>) cache.poll();
-		if (ref != null) {
-			Fragment frag = ref.get();
-			if (frag != null) {
-				return ref.get();
-			}
-		}
-		
-		// this must have been null, remove it
-		cache.remove(ref);
-		return null;
+		return ref != null ? ref.get() : null;
 	}
 
 	/**
@@ -62,6 +54,15 @@ public class AvailableFragmentCache {
 	private void construct(Fragment fragment) {
 		for (FragmentConstructor constructor : constructors) {
 			constructor.construct(fragment);
+		}
+	}
+
+	public void clean() {
+		Iterator<SoftReference<Fragment>> fragRefIterator = cache.iterator();
+		for (SoftReference<Fragment> fragRef : (Iterable<SoftReference<Fragment>>) () -> fragRefIterator) {
+			if (fragRef == null || fragRef.get() == null) {
+				fragRefIterator.remove();
+			}
 		}
 	}
